refactor: 1~2단계 요구사항이 아닌 코드 제거
