--- src/main/java/roomescape/domain/Waiting.java
@@ -1,46 +1,0 @@
-package roomescape.domain;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.OneToOne;
-import java.time.LocalDateTime;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-@EntityListeners(AuditingEntityListener.class)
-@Entity
-public class Waiting {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    @CreatedDate
-    @Column(nullable = false, updatable = false)
-    private LocalDateTime savedDateTime;
-    @Enumerated(EnumType.STRING)
-    private ReservationStatus status;
-
-    @OneToOne(mappedBy = "waiting")
-    private Reservation reservation;
-
-    public Waiting() {
-    }
-
-    public Waiting(ReservationStatus status) {
-        this.status = status;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public ReservationStatus getStatus() {
-        return status;
-    }
-}
