+++ src/main/resources/data.sql
@@ -21,32 +21,17 @@ VALUES ('어드민', 'admin@email.com', 'password', 'ADMIN'),
        ('브리', 'brie@email.com', 'brie', 'USER'),
        ('솔라', 'solar@email.com', 'solar', 'USER');
 
-INSERT INTO waiting (saved_date_time, status)
-VALUES ('2025-01-01 10:00', 'RESERVED'),
-       ('2025-01-01 10:00', 'RESERVED'),
-       ('2025-01-01 10:00', 'RESERVED'),
-       ('2025-01-01 10:00', 'RESERVED'),
-       ('2025-01-01 10:00', 'RESERVED'),
-       ('2025-01-01 10:00', 'RESERVED'),
-       ('2025-01-01 10:00', 'RESERVED'),
-       ('2025-01-01 10:00', 'RESERVED'),
-       ('2025-01-01 10:00', 'RESERVED'),
-       ('2025-01-01 10:00', 'RESERVED'),
-       ('2025-01-01 10:00', 'RESERVED'),
-       ('2025-01-01 10:00', 'RESERVED'),
-       ('2025-01-01 10:00', 'RESERVED');
-
-INSERT INTO reservation (member_id, theme_id, date, time_id, waiting_id)
-VALUES (2, 1, '2025-05-07', 1, 1),
-       (4, 1, '2025-05-08', 2, 2),
-       (3, 1, '2025-05-09', 1, 3),
-       (3, 1, '2025-05-11', 1, 4),
-       (3, 2, '2025-05-11', 1, 5),
-       (3, 2, '2025-05-12', 1, 6),
-       (3, 2, '2025-05-12', 1, 7),
-       (3, 8, '2025-05-12', 1, 8),
-       (3, 8, '2025-05-12', 1, 9),
-       (3, 8, '2025-05-12', 1, 10),
-       (3, 9, '2025-05-12', 1, 11),
-       (3, 11, '2025-05-12', 1, 12),
-       (1, 11, '2025-05-12', 1, 13);
+INSERT INTO reservation (member_id, theme_id, date, time_id)
+VALUES (2, 1, '2025-05-07', 1),
+       (4, 1, '2025-05-08', 2),
+       (3, 1, '2025-05-09', 1),
+       (3, 1, '2025-05-11', 1),
+       (3, 2, '2025-05-11', 1),
+       (3, 2, '2025-05-12', 1),
+       (3, 2, '2025-05-12', 1),
+       (3, 8, '2025-05-12', 1),
+       (3, 8, '2025-05-12', 1),
+       (3, 8, '2025-05-12', 1),
+       (3, 9, '2025-05-12', 1),
+       (3, 11, '2025-05-12', 1),
+       (1, 11, '2025-05-12', 1);
